requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests==2.28.2
python-dotenv==1.0.0
minio==7.1.14
apache-airflow-providers-docker==3.7.1
sqlalchemy

